+++ library/src/jvmMain/kotlin/dev/nesk/akkurate/constraints/ConstraintRegistry.kt
@@ -19,12 +19,27 @@ package dev.nesk.akkurate.constraints
 
 import dev.nesk.akkurate.Configuration
 import dev.nesk.akkurate.ValidationResult
+import kotlin.contracts.ExperimentalContracts
+import kotlin.contracts.InvocationKind
+import kotlin.contracts.contract
 
 /**
  * A storage for all the unsatisfied constraints encountered during a validation.
  */
-internal class ConstraintRegistry {
+internal class ConstraintRegistry(private val configuration: Configuration) {
     private val constraints = mutableSetOf<ConstraintDescriptor>()
+
+    /**
+     * Runs all the checks for [Configuration.failOnFirstViolation]. Should be called between each constraint registration.
+     *
+     * @throws FirstViolationException If an unsatisfied constraint is registered and [Configuration.failOnFirstViolation] is `true`.
+     */
+    fun checkFirstViolationConfiguration() {
+        if (!configuration.failOnFirstViolation) return
+        if (constraints.isEmpty()) return
+
+        throw FirstViolationException(constraints.first().toConstraintViolation(configuration))
+    }
 
     /**
      * Registers the provided constraint if it's unsatisfied.
@@ -51,9 +66,11 @@ internal class ConstraintRegistry {
  */
 internal inline fun <T> runWithConstraintRegistry(
     value: T, configuration: Configuration, block: (ConstraintRegistry) -> Unit,
-) = ConstraintRegistry()
-    .also(block)
-    .toSet()
+) = ConstraintRegistry(configuration)
+    .alsoCatching<ConstraintRegistry, FirstViolationException>(block)
+    .map { it.toSet() }
+    .recover { setOf((it as FirstViolationException).violation) }
+    .getOrThrow()
     .takeIf { it.isNotEmpty() }
     ?.let { ValidationResult.Failure(it.toViolationSet(configuration)) }
     ?: ValidationResult.Success(value)
@@ -63,8 +80,35 @@ private fun Iterable<ConstraintDescriptor>.toViolationSet(configuration: Configu
         .toSet()
         .let(::ConstraintViolationSet)
 
+/**
+ * Calls the specified function [block] with `this` value as its argument and returns a result.
+ *
+ * @throws Throwable If an exception has been throw and doesn't match [E].
+ * @return A [successful result][Result] if no exception happened, a [failed result][Result] if an exception matching [E] has been thrown.
+ */
+@OptIn(ExperimentalContracts::class)
+private inline fun <T, reified E : Throwable> T.alsoCatching(block: (T) -> Unit): Result<T> {
+    contract {
+        callsInPlace(block, InvocationKind.EXACTLY_ONCE)
+    }
+    return try {
+        block(this)
+        Result.success(this)
+    } catch (e: Throwable) {
+        if (e !is E) {
+            throw e
+        }
+        Result.failure(e)
+    }
+}
+
 // TODO: Benchmark `is` usage in all runtimes, compared to the visitor pattern.
 private fun ConstraintDescriptor.toConstraintViolation(configuration: Configuration) = when (this) {
     is Constraint -> toConstraintViolation(configuration.defaultViolationMessage, configuration.rootPath)
     is ConstraintViolation -> this
 }
+
+/**
+ * An internal exception used in conjunction with [Configuration.failOnFirstViolation].
+ */
+internal class FirstViolationException(val violation: ConstraintViolation) : RuntimeException()